+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/dsl/Patchable.java
@@ -23,10 +23,12 @@ public interface Patchable<T> {
   /**
    * Update field(s) of a resource using a JSON patch.
    *
-   * <br>It is the same as calling {@link #patch(PatchContext, Object)} with {@link PatchType#JSON} specified.
+   * <br>
+   * It is the same as calling {@link #patch(PatchContext, Object)} with {@link PatchType#JSON} specified.
    *
-   * WARNING: This may overwrite concurrent changes (between when you obtained your item and the current state) in an unexpected way.
-   * Consider using edit instead.
+   * WARNING: This may overwrite concurrent changes (between when you obtained your item and the current state) in an unexpected
+   * way.
+   * Consider using edit instead or ensure you have called load or withItem to define the base of your patch
    *
    * @param item to be patched with patched values
    * @return returns deserialized version of api server response
@@ -39,13 +41,14 @@ public interface Patchable<T> {
    * Update field(s) of a resource using type specified in {@link PatchContext}(defaults to strategic merge if not specified).
    *
    * <ul>
-   * <li>{@link PatchType#JSON} - will create a JSON patch against the current item.
-   * WARNING: This may overwrite concurrent changes (between when you obtained your item and the current state) in an unexpected way.
-   * Consider using edit instead.
+   * <li>{@link PatchType#JSON} - will create a JSON patch against the current item. See the note in {@link #patch(Object)}
+   * about what is used for the base object.
    * <li>{@link PatchType#JSON_MERGE} - will send the serialization of the item as a JSON MERGE patch.
    * Set the resourceVersion to null to prevent optimistic locking.
    * <li>{@link PatchType#STRATEGIC_MERGE} - will send the serialization of the item as a STRATEGIC MERGE patch.
    * Set the resourceVersion to null to prevent optimistic locking.
+   * <li>{@link PatchType#SERVER_SIDE_APPLY} - will send the serialization of the item as a SERVER SIDE APPLY patch.
+   * You may explicitly set the {@link PatchContext#getFieldManager()} as well to override the default.
    * </ul>
    *
    * @param item to be patched with patched values